+++ kotest-assertions/kotest-assertions-core/src/jvmTest/kotlin/com/sksamuel/kotest/assertions/timing/EventuallyTest.kt
@@ -1,10 +1,18 @@
+@file:Suppress("BlockingMethodInNonBlockingContext")
+
 package com.sksamuel.kotest.assertions.timing
 
+import io.kotest.assertions.assertSoftly
 import io.kotest.assertions.fail
 import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.assertions.timing.EventuallyConfig
 import io.kotest.assertions.timing.eventually
+import io.kotest.assertions.until.fibonacci
+import io.kotest.assertions.until.fixed
+import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.ints.shouldBeLessThan
+import io.kotest.matchers.longs.shouldBeGreaterThan
 import io.kotest.matchers.longs.shouldBeGreaterThanOrEqual
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
@@ -12,13 +20,18 @@ import io.kotest.matchers.string.shouldContain
 import kotlinx.coroutines.delay
 import java.io.FileNotFoundException
 import java.io.IOException
-import kotlin.time.ExperimentalTime
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.Executors
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.atomic.AtomicInteger
 import kotlin.time.TimeSource
 import kotlin.time.days
 import kotlin.time.milliseconds
 import kotlin.time.seconds
+import kotlinx.coroutines.asCoroutineDispatcher
+import kotlinx.coroutines.async
+import kotlinx.coroutines.withContext
 
-@OptIn(ExperimentalTime::class)
 class EventuallyTest : WordSpec() {
 
    init {
@@ -28,7 +41,7 @@ class EventuallyTest : WordSpec() {
                System.currentTimeMillis()
             }
          }
-         "pass tests that completed within the time allowed"  {
+         "pass tests that completed within the time allowed" {
             val end = System.currentTimeMillis() + 250
             eventually(1.seconds) {
                if (System.currentTimeMillis() < end)
@@ -64,7 +77,7 @@ class EventuallyTest : WordSpec() {
          }
          "fail tests throw unexpected exception type"  {
             shouldThrow<NullPointerException> {
-               eventually(2.seconds, IOException::class) {
+               eventually(2.seconds, exceptionClass = IOException::class) {
                   (null as String?)!!.length
                }
             }
@@ -106,7 +119,7 @@ class EventuallyTest : WordSpec() {
                   }
                }
             }.message
-            message.shouldContain("Eventually block failed after 100ms; attempted \\d+ time\\(s\\); 25.0ms delay between attempts".toRegex())
+            message.shouldContain("Eventually block failed after 100ms; attempted \\d+ time\\(s\\); FixedInterval\\(duration=25.0ms\\) delay between attempts".toRegex())
             message.shouldContain("The first error was caused by: first")
             message.shouldContain("The last error was caused by: last")
          }
@@ -116,12 +129,43 @@ class EventuallyTest : WordSpec() {
                System.currentTimeMillis()
             }
          }
-         "allow configuring poll delay" {
+         "allow configuring interval delay" {
             var count = 0
-            eventually(200.milliseconds, 40.milliseconds) {
+            eventually(200.milliseconds, 40.milliseconds.fixed()) {
                count += 1
             }
             count.shouldBeLessThan(6)
+         }
+         "do one final iteration if we never executed before interval expired" {
+            val dispatcher = Executors.newSingleThreadExecutor().asCoroutineDispatcher()
+            async(dispatcher) {
+               Thread.sleep(2000)
+            }
+            val counter = AtomicInteger(0)
+            withContext(dispatcher) {
+               // we won't be able to run in here
+               eventually(1.seconds, 100.milliseconds) {
+                  counter.incrementAndGet()
+               }
+            }
+            counter.get().shouldBe(1)
+         }
+         "do one final iteration if we only executed once and the last delay > interval" {
+            val dispatcher = Executors.newSingleThreadExecutor().asCoroutineDispatcher()
+            // this will start immediately, free the dispatcher to allow eventually to run once, then block the thread
+            async(dispatcher) {
+               delay(100.milliseconds)
+               Thread.sleep(500)
+            }
+            val counter = AtomicInteger(0)
+            withContext(dispatcher) {
+               // this will execute once immediately, then the earlier async will steal the thread
+               // and then since the delay has been > interval and times == 1, we will execute once more
+               eventually(250.milliseconds, 25.milliseconds) {
+                  counter.incrementAndGet() shouldBe 2
+               }
+            }
+            counter.get().shouldBe(2)
          }
          "handle shouldNotBeNull" {
             val mark = TimeSource.Monotonic.markNow()
@@ -132,6 +176,131 @@ class EventuallyTest : WordSpec() {
                }
             }
             mark.elapsedNow().toLongMilliseconds().shouldBeGreaterThanOrEqual(50)
+         }
+
+         "eventually with boolean predicate" {
+            eventually(5.seconds) {
+               System.currentTimeMillis() > 0
+            }
+         }
+
+         "eventually with boolean predicate and interval" {
+            eventually(5.seconds, 1.seconds.fixed()) {
+               System.currentTimeMillis() > 0
+            }
+         }
+
+         "eventually with T predicate" {
+            var t = ""
+            eventually(5.seconds, predicate = { t == "xxxx" }) {
+               t += "x"
+            }
+         }
+
+         "eventually with T predicate and interval" {
+            var t = ""
+            val result = eventually(5.seconds, 250.milliseconds.fixed(), predicate = { t == "xxxxxxxxxxx" }) {
+               t += "x"
+               t
+            }
+            result shouldBe "xxxxxxxxxxx"
+         }
+
+         "eventually with T predicate, interval, and listener" {
+            var t = ""
+            val latch = CountDownLatch(5)
+            val result = eventually(
+               5.seconds,
+               250.milliseconds.fixed(),
+               predicate = { t == "xxxxxxxxxxx" },
+               listener = { latch.countDown() },
+            ) {
+               t += "x"
+               t
+            }
+            latch.await(15, TimeUnit.SECONDS) shouldBe true
+            result shouldBe "xxxxxxxxxxx"
+         }
+
+         "fail tests that fail a predicate" {
+            shouldThrow<AssertionError> {
+               eventually(1.seconds, predicate = { it == 2 }) {
+                  1
+               }
+            }
+         }
+
+         "support fibonacci intervals" {
+            var t = ""
+            val latch = CountDownLatch(5)
+            val result = eventually(
+               duration = 10.seconds,
+               interval = 200.milliseconds.fibonacci(),
+               predicate = { t == "xxxxxx" },
+               listener = { latch.countDown() },
+            ) {
+               t += "x"
+               t
+            }
+            latch.await(10, TimeUnit.SECONDS) shouldBe true
+            result shouldBe "xxxxxx"
+         }
+
+         "eventually has a shareable configuration" {
+            val slow = EventuallyConfig(duration = 5.seconds)
+
+            var i = 0
+            val fast = slow.copy(retries = 1)
+
+            assertSoftly {
+               slow.retries shouldBe Int.MAX_VALUE
+               fast.retries shouldBe 1
+               slow.duration shouldBe 5.seconds
+               fast.duration shouldBe 5.seconds
+            }
+
+            eventually(slow) {
+               5
+            }
+
+            eventually(fast, predicate = { i == 1 }) {
+               i++
+            }
+
+            i shouldBe 1
+         }
+
+         "throws if retry limit is exceeded" {
+            val message = shouldThrow<AssertionError> {
+               eventually(EventuallyConfig(retries = 2)) {
+                  1 shouldBe 2
+               }
+            }.message
+
+            message.shouldContain("Eventually block failed after Infinity")
+            message.shouldContain("attempted 2 time(s)")
+         }
+
+         "override assertion to hard assertion before executing assertion and reset it after executing" {
+            val target = System.currentTimeMillis() + 1000
+            val message = shouldThrow<AssertionError> {
+               assertSoftly {
+                  withClue("Eventually which should pass") {
+                     eventually(2.seconds) {
+                        System.currentTimeMillis() shouldBeGreaterThan target
+                     }
+                  }
+                  withClue("1 should never be 2") {
+                     1 shouldBe 2
+                  }
+                  withClue("2 should never be 3") {
+                     2 shouldBe 3
+                  }
+               }
+            }.message
+
+            message shouldContain "1) 1 should never be 2"
+            message shouldContain "2) 2 should never be 3"
          }
       }
    }