+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/NondeterministicHelpers.kt
@@ -1,0 +1,6 @@
+package io.kotest.assertions
+
+typealias SuspendingProducer<T> = suspend () -> T
+
+
+