+++ kotest-assertions/kotest-assertions-shared/src/commonMain/kotlin/io/kotest/assertions/until/FixedInterval.kt
@@ -1,17 +1,16 @@
 package io.kotest.assertions.until
 
 import kotlin.time.Duration
-import kotlin.time.ExperimentalTime
 
 /**
  * Generates a fixed (linear) poll interval based on the supplied duration
  */
-@OptIn(ExperimentalTime::class)
 class FixedInterval(private val duration: Duration) : Interval {
+   override fun toString() = "FixedInterval(${::duration.name}=$duration)"
+
    override fun next(count: Int): Duration {
       return duration
    }
 }
 
-@OptIn(ExperimentalTime::class)
-fun Duration.fixed() = FixedInterval(this)
+fun Duration.fixed(): FixedInterval = FixedInterval(this)