+++ kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/spec/style/scopes/StringSpecRootScope.kt
@@ -3,11 +3,14 @@ package io.kotest.core.spec.style.scopes
 import io.kotest.core.Tag
 import io.kotest.core.extensions.TestCaseExtension
 import io.kotest.core.test.EnabledIf
+import io.kotest.core.test.NestedTest
+import io.kotest.core.test.TestCase
 import io.kotest.core.test.TestCaseSeverityLevel
 import io.kotest.core.test.TestContext
 import io.kotest.core.test.createTestName
+import kotlin.coroutines.CoroutineContext
+import kotlin.jvm.JvmName
 import kotlin.time.Duration
-import kotlin.time.ExperimentalTime
 
 /**
  * Defines the DSL for creating tests in the 'StringSpec' style.
@@ -21,7 +24,6 @@ import kotlin.time.ExperimentalTime
  */
 interface StringSpecRootScope : RootScope {
 
-   @OptIn(ExperimentalTime::class)
    fun String.config(
       enabled: Boolean? = null,
       invocations: Int? = null,
@@ -49,6 +51,28 @@ interface StringSpecRootScope : RootScope {
    /**
     * Adds a String Spec test using the default test case config.
     */
-   operator fun String.invoke(test: suspend TestContext.() -> Unit) =
-      registration().addTest(createTestName(null, this, false), xdisabled = false, test = test)
+   operator fun String.invoke(test: suspend StringSpecScope.() -> Unit) =
+      registration().addTest(
+         createTestName(null, this, false),
+         xdisabled = false,
+         test = { StringSpecScope(this.coroutineContext, testCase).test() }
+      )
+}
+
+/**
+ * This scope exists purely to stop nested string scopes.
+ */
+class StringSpecScope(
+   override val coroutineContext: CoroutineContext,
+   override val testCase: TestCase
+) : TestContext {
+
+   @Deprecated("Cannot nest string scope tests", level = DeprecationLevel.ERROR)
+   @JvmName("nestedStringInvoke")
+   operator fun String.invoke(test: suspend StringSpecScope.() -> Unit) {
+   }
+
+   override suspend fun registerTestCase(nested: NestedTest) {
+      error("Invalid state")
+   }
 }