+++ kotest-framework/kotest-framework-api/src/commonMain/kotlin/io/kotest/core/test/NestedTest.kt
@@ -3,6 +3,7 @@ package io.kotest.core.test
 import io.kotest.core.factory.FactoryId
 import io.kotest.core.SourceRef
 import io.kotest.core.config.configuration
+import io.kotest.core.plan.Descriptor
 import io.kotest.core.spec.Spec
 
 /**
@@ -15,22 +16,25 @@ data class NestedTest(
    val config: TestCaseConfig,
    val type: TestType,
    val sourceRef: SourceRef,
-   val factoryId: FactoryId?
+   val factoryId: FactoryId?,
+   val descriptor: Descriptor.TestDescriptor?,
 )
 
 /**
  * Returns a full [TestCase] from this nested test, attaching the nested test to the given spec.
  */
-fun NestedTest.toTestCase(spec: Spec, parent: Description): TestCase {
+fun NestedTest.toTestCase(spec: Spec, parent: TestCase): TestCase {
    val testCase = TestCase(
-      description = parent.append(this.name, type),
+      description = parent.description.append(this.name, type),
       spec = spec,
       test = test,
       source = sourceRef,
       type = type,
       config = config,
       factoryId = factoryId,
-      assertionMode = null
+      assertionMode = null,
+      descriptor = descriptor,
+      parent = parent,
    )
    return if (configuration.testNameAppendTags) {
       TestCase.appendTagsInDisplayName(testCase)