+++ kotest-runner/kotest-runner-junit5/src/jvmTest/kotlin/com/sksamuel/kotest/runner/junit5/JUnitTestRunnerListenerTest.kt
@@ -15,9 +15,7 @@ import org.junit.platform.engine.TestDescriptor
 import org.junit.platform.engine.TestExecutionResult
 import org.junit.platform.engine.UniqueId
 import org.junit.platform.engine.reporting.ReportEntry
-import kotlin.time.ExperimentalTime
 
-@OptIn(ExperimentalTime::class)
 class JUnitTestRunnerListenerTests : FunSpec({
 
    test("a bad test should fail test but not parent or spec") {
@@ -48,7 +46,8 @@ class JUnitTestRunnerListenerTests : FunSpec({
          JUnitTestRunnerListenerTests(),
          { },
          sourceRef(),
-         TestType.Container
+         TestType.Container,
+         parent = null,
       )
 
       val test2 = TestCase(
@@ -56,7 +55,8 @@ class JUnitTestRunnerListenerTests : FunSpec({
          JUnitTestRunnerListenerTests(),
          { },
          sourceRef(),
-         TestType.Container
+         TestType.Container,
+         parent = null,
       )
 
       val listener = JUnitTestEngineListener(engineListener, root)