+++ kotest-tests/kotest-tests-timeout/src/test/kotlin/com/sksamuel/kotest/timeout/TestCaseTimeoutListenerTest.kt
@@ -17,11 +17,9 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
 import java.util.concurrent.atomic.AtomicInteger
 import kotlin.coroutines.CoroutineContext
-import kotlin.time.ExperimentalTime
 import kotlin.time.milliseconds
 
 @Suppress("BlockingMethodInNonBlockingContext")
-@ExperimentalTime
 class TestCaseTimeoutListenerTest : FunSpec() {
 
    private val blockingCount = AtomicInteger(0)
@@ -45,7 +43,8 @@ class TestCaseTimeoutListenerTest : FunSpec() {
 
          val testCase = TestCase.test(
             TestCaseTimeoutListenerTest::class.toDescription().appendTest("wibble"),
-            this@TestCaseTimeoutListenerTest
+            this@TestCaseTimeoutListenerTest,
+            parent = null,
          ) {
             Thread.sleep(1000000)
          }.copy(
@@ -81,7 +80,8 @@ class TestCaseTimeoutListenerTest : FunSpec() {
 
          val testCase = TestCase.test(
             TestCaseTimeoutListenerTest::class.toDescription().appendTest("wibble"),
-            this@TestCaseTimeoutListenerTest
+            this@TestCaseTimeoutListenerTest,
+            parent = null,
          ) {
             delay(1000000)
          }.copy(